LayoutTests/imported/w3c/web-platform-tests/wpt_internal/html/canvas/drawElementImage/draw-element-image-get-image-data-expected.txt


FAIL canvas drawElementImage readback assert_equals: expected 1 but got 3
